Add macOS ARM64 pipeline for webgpu (#24060)
### Description

Add macOS ARM64 pipeline for webgpu.

This pipeline is a temporary one. I created this pipeline because the
current code already fails on macOS ARM64 for WebGPU EP. Adding this
pipeline allows to check the status of the fix, and eventually when the
build passes, this pipeline will be merged with the existing macOS arm64
pipeline.