Change type len from int to size_t (#24157)
### Description
As titled.



### Motivation and Context
We have the last MatMul in phi-4-mini onnx which is b_shape = {3072,
200064}
packed_b_size = MlasGemmPackBSize(N, K);
 it is `3072*200064*sizeof(float)=2458386432`
This is larger than 2,147,483,647, it is out of the int boundary on a
32-bit system. Then len is negative.
So we change the type to size_t, and the model can be loaded
successfully after the change.