Add shader key validation step in WebGPU CI pipeline (#24243)
### Description

This PR adds a shader key validation step to the WebGPU CI pipeline.

The shader key validation works in this way:

- first, run onnxruntime_test_all with verbose logging, dumping the logs
into a file
- then, parse the file and found WebGPU EP program logs. The log
contains the following information:
  - the shader cache key
  - the corresponding shader code
  
The script will aggregate those information and make sure for each cache
key, the corresponding shader code must be consistent.

To make the validation work, this PR also modified a few things:
- set the locale of `std::wclog` to ".UTF-8" to support Unicode
characters. Otherwise the logger will fail and no longer output future
logs. A fix is submitted in PR #24237 but there is a concern if this may
potentially break some users. Setting inside onnxruntime_test_all is
pretty safe.
- re-enable the WebGPU device auto collect which was introduced in
https://github.com/microsoft/onnxruntime/pull/24115. Now we have a
better way to detect cache key inconsistency.

### Next Step

The newly added test is marked as `continue-on-error: true`, which means
even if it failed it does not block the CI pipeline. We should fix those
failures one-by-one and eventually the test should pass. then we can
remove the `continue-on-error: true` flag.