Update packaging pipeline for Nodejs binding (#24301)
### Description

Update packaging pipeline for Nodejs binding.

This change updates the pipeline to perform all Node.js binding builds,
including:
- Windows x64 ( CPU, DML, WebGPU )
- Windows arm64 ( CPU, DML, WebGPU )
- Linux x64 ( CPU, CUDA, TensorRT, WebGPU )
- Linux arm64 ( CPU )
- MacOS x64 ( CPU, CoreML, WebGPU )
- MacOS arm64 ( CPU, CoreML, WebGPU )

#### Dependencies

The Node.js binding depends on the Nuget package from the same build.

Because NPM has a size limit so we cannot fit
libonnxruntime_provider_cuda.so into it. The Node.js binding works in a
way that an installation script will try to download the Nuget package
of the corresponding version.