[AISW-134328] Check in copy of AI Hub build script
To kick of building a fast and debuggable CI system, we begin by
checking in a stripped-down copy of the top-level script used by
the AI Hub team for both local development and CI.

Please note that script does not yet do anything useful. Tasks
will be added as subsequent items in epic AISW-134319 are completed.