Fix cli in dockerfiles/README.md (#23493)
Added context to command line example when specifying platform.

### Description
Docker image build fails due to missing context in the command line
example when specifying the platform.

The dockerfiles directory is assumed as the current directory in the
command example, so the parent directory must be specified as the
context of the docker build command