[CoreML] Add support for int64 (#24462)
### Description
Add int64 as a supported datatype for moving nodes to the CoreML EP.

We already convert constants automatically from int64 to int32 for
CoreML by calling narrow.

Adding the conversion for outputs as well.

### Motivation and Context
- More nodes supported on CoreML

### Note on the Unsqueeze op
According to #22975 there is a bug with the Unsqueeze op with scalar
inputs on x86.

I was running into a bug for unsqueezes that unsqueezed a scalar input
to a tensor of shape [1] since CoreML doesn't support scalar values for
MLProgram. I adapted the HandleX86ArchUnsqueeze method but
alternatively, can replace with an identity operator or add some
additional checks. I went with adapting the HandleX86ArchUnsqueeze
method since it seemed like the fastest solution.