[NV EP] fix EP context options  (#24545)
### Description

While cleaning up the options I missed the part in the provider bridge
that translates session options to TRT options.
To better integrate with current IHV work I adopted the principle that
QNN and OV use to pipe through session options. Since all this is string
based magic it would be great to be access a general point of truth like
`EpContextModelGenerationOptions` in the provider wrappedtypes.

https://github.com/microsoft/onnxruntime/blob/896600b83be5072548e84614716deae248944233/onnxruntime/core/framework/session_options.h#L73

This is a fix on top of #24456 @ankan-ban and @chilo-ms to review.