[AISW-129779] Define SpaceToDepth fusion for YOLOv2.
A sequence of Reshape and Transpose have 6D shapes and are falling
back to CPU for YOLOv2 model. Add fusion to fuse this sequence of ops
into a single SpaceToDepth. Note that the pattern in YOLOv2 is
different from the common pattern of SpaceToDepth which left as
future work to be implemented.
Since matching the SpaceToDepth pattern requires model with shapes
inferred, add shape inference in `qnn_preprocess_model` before
invoking the fusion.
In addition, `onnxruntime/python/tools/qnn/preprocess.py` is added
to provide an entry point/API for non-quantization user to preprocess
models for QNN. Even though this file in facts invokes the function
under `onnxruntime/python/tools/quantization/`, it is provided to
avoid confusion for user. Furthermore, the directory `tools/qnn` will
be packaged into the wheel for the usage from cmdline.