[VITISAI] pass all session configs to vitisai ep for `Ort::CompileModel` flow (#24799)
### Description

convert all session configs, i.e. key-value pairs into provider options, the key prefixed with `ort_session_config.`


### Motivation and Context
#24445 has a bug when `Ort::CompileModel` is used, not all session config are passed to VITISAI EP backend.
It is because that the `session_option` which holds a reference to `VitisiAIExectuionProviderFactory` is not as same as the
`session_option` used for `Ort::CompileModel`. `Ort::CompileModel` create another `session_option` behind scene.

The symptom of this bug is that only the session configs in the first `SessionOptions` object is passed to `VitisiAIExectuionProviderFactory` and session configs in the second `SessionOptions` are not, so that VITISAI EP backend sometimes assumes that ep.cache_context is not enabled, and then ep context cache model is not created properly.